+++ migrations/2020_01_24_000000_modify_meta_indexes.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class ModifyMetaIndexes extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('meta', function (Blueprint $table) {
+            $table->dropIndex(['metable_type', 'metable_id']);
+            $table->unique(['metable_type', 'metable_id', 'key']);
+            $table->index(['key', 'metable_type']);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('meta', function (Blueprint $table) {
+            $table->dropIndex(['key', 'metable_type']);
+            $table->dropUnique(['metable_type', 'metable_id', 'key']);
+            $table->index(['metable_type', 'metable_id']);
+        });
+    }
+}